+++ climpred/tests/test_hindcastEnsemble_class.py
@@ -4,14 +4,14 @@ from climpred import HindcastEnsemble
 
 
 def test_hindcastEnsemble_init(hind_ds_initialized_1d):
-    """Test to see hindcast ensemble can be initialized"""
+    """Test to see hindcast ensemble can be initialized with xr.Dataset."""
     hindcast = HindcastEnsemble(hind_ds_initialized_1d)
     print(hindcast)
     assert hindcast
 
 
 def test_hindcastEnsemble_init_da(hind_da_initialized_1d):
-    """Test to see hindcast ensemble can be initialized with da"""
+    """Test to see hindcast ensemble can be initialized with xr.DataArray."""
     hindcast = HindcastEnsemble(hind_da_initialized_1d)
     assert hindcast
 
@@ -162,3 +162,13 @@ def test_inplace(
     hindcast.sum('init')
     summed = hindcast.sum('init')
     assert hindcast != summed
+
+
+def test_verify_metric_kwargs(hindcast_hist_obs_1d):
+    """Test that HindcastEnsemble works with metrics using metric_kwargs."""
+    assert hindcast_hist_obs_1d.verify(
+        metric='threshold_brier_score',
+        comparison='m2o',
+        threshold=0.5,
+        reference='historical',
+    )